refactor: modals
