+++ src/app/modals/newsletter/newsletter.module.ts
@@ -1,0 +1,13 @@
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+
+import {IonicModule} from '@ionic/angular';
+
+import {NewsletterPage} from './newsletter.page';
+
+@NgModule({
+  imports: [CommonModule, FormsModule, IonicModule, ReactiveFormsModule],
+  declarations: [NewsletterPage],
+})
+export class NewsletterPageModule {}
