+++ src/app/modals/newsletter/newsletter.page.scss
