--- src/app/pages/feedback/feedback-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { FeedbackPage } from './feedback.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: FeedbackPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class FeedbackPageRoutingModule {}
