+++ src/app/pages/home/home.module.ts
@@ -2,12 +2,12 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {IonicModule} from '@ionic/angular';
 import {FormsModule} from '@angular/forms';
+import {ReactiveFormsModule} from '@angular/forms';
+
 import {HomePage} from './home.page';
 
 import {HomePageRoutingModule} from './home-routing.module';
 import {ComponentsModule} from 'src/app/components/components.module';
-
-import { ReactiveFormsModule } from "@angular/forms";
 
 @NgModule({
   imports: [CommonModule, FormsModule, IonicModule, HomePageRoutingModule, ComponentsModule, ReactiveFormsModule],
