--- src/app/pages/newsletter/newsletter-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { NewsletterPage } from './newsletter.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: NewsletterPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class NewsletterPageRoutingModule {}
