--- src/app/pages/newsletter/newsletter.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule,ReactiveFormsModule } from '@angular/forms';
-
-import { IonicModule } from '@ionic/angular';
-
-import { NewsletterPageRoutingModule } from './newsletter-routing.module';
-
-import { NewsletterPage } from './newsletter.page';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    NewsletterPageRoutingModule,
-    ReactiveFormsModule
-  ],
-  declarations: [NewsletterPage]
-})
-export class NewsletterPageModule {}
